Actually implement timer behavior
This includes syncing with indoor unit, offline ticking, and persisting the states.

Currently it's unclear how the indoor unit will behave when both ON and OFF timers are active (official remote never sets both at the same time). Assuming that both will be ticking at the same time for now. More experimenting is needed on this.