Fix some buffer size issues, preventing crashes
This commit fixes some issues with buffers, reducing the amount of crashes on extreme inputs.

1. Consistently define buffer sizes to include room for check digit and NULL-terminator
2. In regards to 1. also remove the additional `+1` from buffer allocations
3. Use safer `strlcpy` to copy NULL-terminated strings
4. Use destination size in `strlcpy` to prevent buffer overflows

The size of `char passport_number[]` in `passy_reader.c` was already limited to `11`. `20` is closer to the previously, inconsistently used `11` and still leaves some extra room aboce the usual `9` characters of a document number.

If, due to an error or extensive input `passy->passport_number` contains more characters than fit in `passport_number` locally, using `strlen` on the source creates a buffer overflow. This is mitigated by using the destination buffer size.