Use proper buffer for progress update
The `char header[32]` is limited to the scope of the `else { ... }`.
Anything written to it might be, and is, overwritten after this block.
Passing this as an argument to `popup_set_header` is therefore unsafe.

With this PR the text_store in `struct Passy` is used, which remains
available and solves the flickering display of progress.